requirements.txt
notebook
ipywidgets
tqdm
python-dotenv
openai
streamlit
rembg
onnxruntime
pillow
torch
transformers
requests